Lots of updates to both classes.
==ShippingProgramGUI==
-Added the site ComboBox, pops up once a customer is selected in the
first CB. Still need to figure out how to not create a new one when the
selection changes.

-Moved all code relating to the creation of CBs into their own methods.
Constructor now simply creates the application frame and calls
"createCustomersBox".

-Added an event listener to the customers CB, when the box's state is
changed the selection is sent to the "createSitesBox", which then
queries the relevant sites from the DBM and creates a new CB.


==DatabaseManager==
-Moved all code relating to querying and populating lists into their own
methods. Constructor now solely registers the driver and calls the
"getCustList" method.

-As mentioned, created "getCustList", "getSiteList", and
"getAccountList" methods. Pretty much what they say on the tin.

-A ton of comments. Still need to do that for the GUI class.