Fairly minor updates to everything.
==DatabaseManager==
-Added getters and setters for the active selection variables.

-Added a method to append insert statements to the
"ShipmentProgramInsert-Customer.sql" file when a new customer is
created. Eventually, a user will be able to add a customer from the GUI.
Under the hood, that process will take the user's input and add it to
the database, next calling this method in order to maintain a backup
file.

==ShippingProgramGUI==
-Created a repopSitesBox method to deal with situations where the user
changes their customer selection. Before, this would cause new boxes to
be created, the last of which would contain the old sites AND the new
sites. Now the box is flushed and repopulated. This works well, so i
will take the same approach when dealing with the accounts box.

-Created a repopAccountsBox method: see above.

-As always, tidied up and added comments.

==SQL files==
-Cleaned up the Customer and Site insert files to match my manual
revisions (typo corrections, etc.). Now everything in these files
matches the actual database contents, meaning I could destroy the
database and launch these files and no one would be the wiser.

-Renamed the folder containing them from "misc" to "sql".