Lots of under-the-hood stuff.
==DatabaseManager==
--Huge strides towards modularization.

-Rather than having three different list getters that do very similar
tasks, there is now one generic "getPrimaryKeysList" that allows a
specific table to be passed as a parameter. The method performs special
checks depending on the user's choice.

-Reduced from three distinct setters and getters down to one that
accepts a tableName parameter.

-To make all of this possible, the different table lists and active
selections were converted from three separate variables to one single
hashmap, where the name of the table serves as the key. This means that
I can use one method to refer to each of the three table lists just by
getting the value of the table name from the map.

-Rather than reusing the code to execute a generic query multiple times,
wrote a helper function called "runBasicQuery".

-Still on the to-do list following in today's footsteps are
	: create a "runExtractQuery" method to trim down the list getter
method's size
	: create methods to update global HashMaps (if needed, I need to run
some tests first)
	: clean up some of the list getter's logic
	: remove the now-deprecated methods and variables.
	: either enumerate or globally store the tableName strings, so there
are no problems with formatting (maybe switch to numbers? then I could
even do mathematical comparisons for some of the validation)

--A couple of general purpose method additions
-"addNewCustomer" and "addNewSite" methods. These accept the necessary
args from the user and add the info to the database and the SQL file
backups. These were the inspiration for the "runBasicQuery" method, as
they were using identical database-related code.

--Tons of reformatting and tidying up, still more to be done.

==ShippingProgramGUI==
-Minor tweaks to play nicely with the new generalized methods.