Some more under the hood changes.
==DatabaseManager==
-Removed deprecated getters, setters, and variables. Now 100% on the
generalized methods.

-Moved from a string literal for tableName arguments to an enum system.

-Toyed around a little with redoing the invalid selection logic, didn't
finish so reverted those changes for now. I'd like to do some ordinal
comparisons but if that ends up being more verbose than my current
solution I'll just skip that.

-Created a "runSelectQuery" method to declutter the "getPrimaryKeyList"
method. Accepts the tableKeys and the name of the field to retrieve.
Ultimately I would like to make this work for the full data retrieval as
well, not just the primary key retrieval. Probably an array-centric
solution for that.

==ShippingProgramGUI==
-Minor edits to play nicely with the enum and other generalization
tweaks made in the DBM.