Some major restructuring. Not much in the way of new features.
==ShippingProgram==
-That's right, a new class. This will sort of act as a central hub for
the program. Currently holds the constants used by the other two
classes, as well as the main method that gets everything kicked off.

-Changed the tableName enum to a set of constant ints, allowing for
mathematical comparisons.

==ShippingProgramGUI==
-No longer a subclass of JFrame. Holds its own instance of JFrame
instead. This was done in an attempt to fix the card layout, as all
tutorials I have seen use this approach. If it still doesn't work, I'll
probably change it back.

-Flattened everything down into universal methods. Replaced the three
create methods with createComboBox, which accepts a tableName. Took the
same approach for the repop methods, as well as the promptNew methods.

-To facilitate this, condensed the three separate JComboBox variables
into a HashMap.

-Old code is just commented out for now, just in case there were any
snippets that didn't properly translate. Will probably remove next time
I work on this.

*Currently a null pointer exception occurs in relation to the creation
of the first comboBox. I will be investigating this next time I work on
this project. Not sure exactly what's going on.