Loads of really great front-end changes, especially pertaining to layout and resizing.
==ShippingProgramGUI==
-Added a controls box, and spent a long time tweaking the layout. Now
things look pretty good, and resizing doesn't make everything look
terrible.

-Added a new contentPane variable, and set the contentPane layout to
BorderLayout.

-Created a new controlPanel JPanel for holding the previous and next
buttons. These will gain functionality soon, and will allow the user to
switch between cards.

-Added three new "create" methods for different GUI elements, to remove
some clutter from the constructor.

-createCardDeck: creates the cardDeck JPanel, sets it to CardLayout,
adds a border, and adds the deck to the contentPane at the NORTH
location.

-createControlPanel: creates the controlPanel JPanel, creates and adds
two buttons, and adds the panel to the contentPane at the SOUTH
location.

-createFirstPage: creates the startPageCard JPanel, sets it to
GridLayout with 3 rows, 1 column, 0 horizontal spacing, and 40 vertical
spacing. Creates the three comboBoxes, disables the SITES and ACCOUNTS
boxes, and adds the startPageCard to the cardDeck (currently I'm
accidentally doing this twice, will fix later)

-repopComboBox now makes sure that the box being repopulated is enabled.

-ComboBoxListener's "itemStateChanged" now repops a box no matter what,
and never calls createComboBox. This is because comboBoxes now exist
from the beginning of runtime. Also checks to see if CUSTOMERS is the
box being changed, and if so, the ACCOUNTS comboBox is disabled.

==DatabaseManager==
-getPrimaryKeyList now returns the basic list ("--Select one--" and
"--Add new--") if invalid selections are made, instead of an empty list.

==ShippingProgram==
-Added code to use the System look and feel for the GUI elements. I may
change this later, I've gotten used to the default Metal look.

-Uncommented threading code.