Some pretty major progress.
==ShippingProgramGUI==
-Made a lot of changes that culminate to a working paging system. The
program now has working next and previous buttons, and can switch
between cards properly. Next up is to figure out what the second page is
gonna be and get it up and running. First page is basically done at this
point.

-Added some global variables to hold a list of the program's cards, the
current card, the active CardLayout manager, and the names of each card.

-Constructor now creates the first and second pages. Second page is a
placeholder.

-createFirstPage now adds the new card to the list, rather than storing
it in a discrete global variable. Also names the card "Page 1" when
adding it to the deck.

-createSecondPage now exists, creates a placeholder second page to test
paging.

-createControlPanel now attaches listeners to the previous and next
buttons. Buttons come complete with logic to enable and disable
themselves based on page number.

-Added a case in promptNewEntry to add a new account entry.

==DatabaseManager==
-Created an addNewAccount method that adds a new account entry to the
database and the relevant SQL file.

-