Lots of reorganizing and future preparation.
==ShippingProgramGUI==
-Cleanup and reorganization.

==DatabaseManager==
-Began preparing for storage/use of full recipient data.

-Created an "activeData" HashMap. This stores an Integer (global
tableName value) as the key, and an ArrayList as the value. The ALists
hold the data for the active selection, with each entry in the list
corresponding to a database column.

-Created a "getCompleteData" method. This populates the "activeData"
HashMap, then returns the data for the given tableName argument as an
ArrayList of Strings.

-Created a method stub for "updateSelectedCustomer". I didn't feel like
starting this today, but it will be done soon. Will allow for updates to
database entries, will be accompanied by similar methods for the other
two tables, as well as GUI methods for popup window prompts.

-Created a second version of "runSelectQuery". The original version has
changed its arg order from "aList, fieldName, sql" to "aList, sql,
fieldName". This version is meant to be used when the aList will be
filled with values of one single column. A second version can be called
with args "aList, sql". This version is to be used when all results from
the sql statement are to be dumped into the list. The first is currently
used to retrieve overall primary key lists, while the second is used to
retrieve full data for one table entry.

-Created a "fetchCompleteData" method to retrieve all data for the
active selections. This will only be called after a selection in each
category has been made.

==Recipient==
-Yes, another new class. This is tentative, I may change my mind on this
implementation. Currently, my plan is to have this be the culmination of
all info from all three tables.

-Created a HashMap called "data" that stores the information aggregated
from the three tables. Keys will be column names pulled from the
database itself, values will be the corresponding entries in the
database for the user's selected recipient.

-"getField" and "setField" that deal with the data HashMap when given a
fieldName (key).

-Ideally, once this is properly implemented, it will represent the
active selection of the user. Data will be fed to it via setters, and
retrieved via getters to complete shipments. All "active selection" data
will be stored here instead of in the DBM.