Mostly some work on the Recipient class and some of the tricky logic surrounding it.
==DatabaseManager==
-Removed the getCompleteData method. It just didn't make much sense in
light of the new Recipient object.

-Repurposed fetchCompleteData to perform a similar task to
getCompleteData. It's now public, is called when the page is advanced,
and feeds its info to the constructor for a Recipient that is stored
globally.

-Moved the activeData HashMap into the fetchCompleteData method, instead
of keeping it global. The global ActiveRecipient will handle the
functionality it was originally intended for.

-Fixed a bug in one of the runSelectQuery methods where the cursor was
not being placed at the beginning of the ResultSet.

==Recipient==
-Fully ditched the HashMap style of keeping up with information.

-Constructor accepts three ALists containing info from each of the three
tables.

-setCustValues, setSiteValues, and setAcctValues take the info from the
ALists and distribute it to the proper attribute variables. Assumption
is made that the ALists are not malformed.

==StarshipGUI==
-Next button listener now calls the dbm's fetchCompleteData method,
creating the ActiveRecipient. This will be tweaked in the future- the
logic is proving a little tricky to wrap my head around. The second page
is created at runtime, but the Recipient data is not available until the
first page is completed. So if I want to display the information
collected on the first page, I'll have to implement some way of
populating the second page upon switching to it. Then again, this may be
an issue with the entire program, so I might just rethink how I create
and populate the pages in general. Gonna sit on this one for a bit.