A little toying around with layouts and such.
==DatabaseManager==
-Changed fetchCompleteData again. Now accepts tableName, and does not
retrieve all three tables' data. Called when a selection is made, and
gives the info to the Recipient object. This is done so info can be
shown as selections are made.

==StarshipGUI==
-No longer fetches all info on "next" click.

-Added three new panels, one for the info from each table. Currently
these are displayed in-line with the ComboBoxes, but this will change
because the layout isn't what I wanted. I'll have to figure out a way to
split the first page into two panes, with the left being the selection
screen and the right containing the three info readout panes, along with
"edit" buttons.

==Recipient==
-Made the info setters public, so they can be called when selections are
made.