Began the process of switching from MySql to SQLite.
==DatabaseManager==
-Removed USERNAME and PASSWORD fields. SQLite does not have user
control.

-DB_URL now points to the file "starship.db" instead of a locally hosted
database. SQLite stores its database on a file, rather than requiring it
to be hosted.

-Added a new method, "checkDatabase". This attempts to connect to the
database and see if it has any data stored. If the file does not exist
at all, the act of attempting a connection will create it. This is
great, because I've been worried about how deploying a new installation
of this program would work. I would've either had to host the database
perpetually and remotely, or figure out a way to create and host it on
program start. Having a file-based solution is much simpler. If the
database is found to contain no data, the method "buildDatabase" will be
called to rebuild the data from the bundled SQL files.

-Created a new method (skeleton), buildDatabase. This is called when
checkDatabase finds no data in the database file. This will eventually
parse all bundled SQL files and run their commands, which will rebuild
the data previously stored. This is good for recovering from loss of the
DB file, but I'm gonna have to figure out a way to save these SQL files
outside of the program in the event of the whole program being lost-
otherwise, it doesn't do much good.

-Removed all references to USERNAME and PASSWORD field in database
accessor methods.

==Misc changes==
-Added the SQLite JDBC driver to the project and classpath.

-Deleted the "StarshipInsert-Accounts" file and fixed the line in
"addToInsertFiles" that accidentally created it.

-Created a DB file for the program. Seems like this will stay bundled
with the program for now, not sure if this is how I'm always gonna
handle it. If I do keep it bundled, the database won't have to be
rebuilt every time the program is run for the first time on a new
machine, which would be a plus.